feat: add Set interface and impl.
This change adds a collections directory that is not intended to be used
as a package itself, but to hold packages that implement common
collection-based data structures. The first package defined under this
directory is `set`, which implements a basic set data structure using a
map where the keys are of generic type T, and the value is an empty
struct.

The `collect/set` package defines a `Set` interface as well as an
unexported type `set` that implements it. New instances of this type may
be created using the exported `New` function. This approach was chosen
in order to have more control over the implementation details of the
`set` implementation. E.g. if I ever decide to make the type a custom
struct that holds a map, rather than using a type definition to make the
`set` a `map`, the zero value of the struct would no longer be useful.
Thus, I decided not to allow composite-literal instantiation of the
`set` type.

The method names were chosen by looking at several other languages' set
implementations and evaluating what seemed most intuitive. E.g. for the
function that adds a new item to the set, I chose to name this "Add", as
Python does, because "Insert" and "Append" both imply adding the element
at a given position.