feat: add Update() and Seq() methods to Set.
Update() is similar to Python's set.update() method. It allows the
caller to pass several sets, and all items from each of the sets are
added to the set on which the method is being called.

Seq() provides an iterator over the set. It's a thin wrapper around
maps.Keys().