refactor: move p/c/set to p/c/sets/hashset
This makes more sense with regard to the way godocs are generated (i.e.
attached to the method on the struct where it's being implemented), and
also is a more intuitive package structure if we want to add more types
of sets (e.g. a TreeSet to take advantage of ordering characteristics).